+++ lib/blocks/blockBase.rb
@@ -212,7 +212,7 @@ module Bake
             if (@tcs[:OUTPUT_DIR] == nil) && (Bake.options.buildDirDelimiter == "/") # in this case all builds are placed in a "build" folder
               buildDir = File.dirname(@output_dir)
               if (File.basename(buildDir) == "build") && (Dir.entries(buildDir).size == 2)# double check if it's really "build" and check if it's empty (except "." and "..")
-                puts "Deleting folder #{@buildDir}" if Bake.options.verbose >= 2
+                puts "Deleting folder #{buildDir}" if Bake.options.verbose >= 2
                 FileUtils.rm_rf(buildDir)
               end
             end