+++ lib/blocks/library.rb
@@ -41,7 +41,7 @@ module Bake
         aTime = File.mtime(@archive_name)
 
         # sources
-        @compileBlock.objects.each do |obj|
+        @objects.each do |obj|
           return "because object #{obj} does not exist" if not File.exists?(obj)
           return "because object #{obj} is newer than executable" if aTime < File.mtime(obj)
         end
@@ -52,9 +52,21 @@ module Bake
       def execute
 
         Dir.chdir(@projectDir) do
-          if @compileBlock.objects.empty?
-            puts "No source files, library won't be created" if Bake.options.verbose >= 2
-            return true
+
+          @objects = @compileBlock.objects
+          if !@block.prebuild
+            if @objects.empty?
+              puts "No source files, library won't be created" if Bake.options.verbose >= 2
+              return true
+            end
+          else
+            @objects = Dir.glob("#{@output_dir}/**/*.o")
+            if @objects.empty?
+              if !File.exists?(@archive_name)
+                puts "No object files, library won't be created" if Bake.options.verbose >= 2
+              end
+              return true
+            end
           end
 
           cmdLineCheck = false
@@ -78,7 +90,7 @@ module Bake
             cmd[cmd.length-1] += @archive_name
           end
 
-          cmd += @compileBlock.objects
+          cmd += @objects
 
           if cmdLineCheck and BlockBase.isCmdLineEqual?(cmd, cmdLineFile)
             success = true
@@ -99,11 +111,19 @@ module Bake
       end
 
       def clean
-        return cleanProjectDir()
+        if @block.prebuild
+          Dir.chdir(@projectDir) do
+            @objects = Dir.glob("#{@output_dir}/**/*.o")
+            if !@objects.empty? && File.exist?(@archive_name)
+              puts "Deleting file #{@archive_name}" if Bake.options.verbose >= 2
+              FileUtils.rm_rf(@archive_name)
+            end
+          end
+        else
+          return cleanProjectDir()
+        end
       end
 
-
     end
-
   end
 end